Fix efi memory allocator.
Now, the allocator does not allocate a whole page each time. 8 pages are
allocated when the program starts, and this buffer is used to allocate
further memory using this buffer.

Update output buffers to be allocated only once.

These two output buffers (one for 8bit chars and one for 16bit chars)
are stored as global pointers.